moved some methods around
